+++ src/test/java/de/stylelabor/statusplugin/util/ColorUtilTest.java
@@ -104,12 +104,12 @@ public class ColorUtilTest {
 
     @Test
     public void testVersionComparison() {
-        assertTrue(VersionChecker.isNewer("7.1", "7.1.1"));
-        assertTrue(VersionChecker.isNewer("7.1.1", "7.2"));
-        assertTrue(VersionChecker.isNewer("7.1.1", "8.0.0"));
-        assertFalse(VersionChecker.isNewer("7.1.1", "7.1.1"));
-        assertFalse(VersionChecker.isNewer("7.1.1", "7.1"));
-        assertFalse(VersionChecker.isNewer("8.0.0", "7.1.1"));
+        assertTrue(VersionChecker.isNewer("7.1.1", "7.1.2"));
+        assertTrue(VersionChecker.isNewer("7.1.2", "7.2"));
+        assertTrue(VersionChecker.isNewer("7.1.2", "8.0.0"));
+        assertFalse(VersionChecker.isNewer("7.1.2", "7.1.2"));
+        assertFalse(VersionChecker.isNewer("7.1.2", "7.1.1"));
+        assertFalse(VersionChecker.isNewer("8.0.0", "7.1.2"));
     }
 
     @Test